# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  inventory.CreateInventoryItemRequest:
    properties:
      category:
        maxLength: 100
        type: string
      expiry_date:
        type: string
      food_item_id:
        type: string
      location:
        maxLength: 100
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      quantity:
        type: number
      unit:
        maxLength: 50
        type: string
    required:
    - name
    - quantity
    type: object
...
  inventory.InventoryItem:
    properties:
      category:
        type: string
      created_at:
        type: string
      expiry_date:
        type: string
      food_item_id:
        type: string
      id:
        type: string
      location:
        type: string
      name:
        type: string
      quantity:
        type: number
      unit:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  inventory.UpdateInventoryItemRequest:
    properties:
      category:
        maxLength: 100
        type: string
      expiry_date:
        type: string
      food_item_id:
        type: string
      location:
        maxLength: 100
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      quantity:
        type: number
      unit:
        maxLength: 50
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go backend API server for Foodlink application
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Foodlink Backend API
  version: "1.0"
paths:
...
  /inventory:
    get:
      consumes:
      - application/json
      description: Get all inventory items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.InventoryItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get user inventory
      tags:
      - inventory
    post:
      consumes:
      - application/json
      description: Add a new item to user's inventory
      parameters:
      - description: Inventory item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inventory.CreateInventoryItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.InventoryItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Add inventory item
      tags:
      - inventory
  /inventory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an inventory item
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Delete inventory item
      tags:
      - inventory
    get:
      consumes:
      - application/json
      description: Get details of a specific inventory item
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.InventoryItem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get inventory item by ID
      tags:
      - inventory
    put:
      consumes:
      - application/json
      description: Update an existing inventory item
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Inventory item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inventory.UpdateInventoryItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.InventoryItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Update inventory item
      tags:
      - inventory
  /inventory/expired:
    get:
      consumes:
      - application/json
      description: Get all expired inventory items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.InventoryItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get expired items
      tags:
      - inventory
  /inventory/expiring:
    get:
      consumes:
      - application/json
      description: Get inventory items expiring within specified days (default 7)
      parameters:
      - description: 'Number of days (default: 7)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.InventoryItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get expiring items
      tags:
      - inventory
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"